feat: add kustomize component to wait for 'alembic' migrations
It uses a simple shell script to make deployments marked with the
label `wait-for-migrations: required` wait until 'alembic' reports
that the (known) migrations have reached the "head", i.e. until
a successful 'alembic' run has finished (e.g. in some other job).

It assumes the 'alembic' binary to be installed via 'pipenv' and
can be added to a manifest like this:

  components:
  - github.com/ZeitOnline/kustomize/components/postgrest?ref=1.4